Add benches for refund relayer extension
